sakila-data-lake/requirements.txt
# Core Dependencies
kafka-python==2.0.2
pyspark==3.5.0

# AWS/S3 Support (compatible versions)
boto3==1.34.1
botocore==1.34.1
s3fs==2024.2.0
minio==7.2.0

# Database Drivers
pymysql==1.1.0
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Configuration
pyyaml==6.0.1

# Monitoring & Logging
python-json-logger==2.0.7

# Development Tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Hadoop AWS (for Spark S3 integration)
# Note: These need to match your Spark version
# hadoop-aws==3.3.4
# aws-java-sdk-bundle==1.12.262